split criteria and tasks in different files
